test: Re-organize tests for `term_image.image`
- Change: Collect all tests for `.image` in one directory.
- Change: Update make targets for tests.
