+++ tests/test_image/test_base.py
