+++ tests/test_iterator.py
