+++ tests/test_top_level.py
@@ -5,40 +5,8 @@ import pytest
 
 from term_image import AutoCellRatio, get_cell_ratio, set_cell_ratio
 from term_image.exceptions import TermImageError
-from term_image.image import AutoImage, BaseImage, ImageSource, from_file
 
 from . import set_cell_size
-from .test_base import python_image, python_img
-
-
-class TestConvinience:
-    def test_auto_image(self):
-        with pytest.raises(TypeError, match=r"'PIL\.Image\.Image' instance"):
-            AutoImage(python_image)
-
-        # Ensure size arguments get through
-        with pytest.raises(ValueError, match=r"both width and height"):
-            AutoImage(python_img, width=1, height=1)
-
-        # Ensure scale argument gets through
-        with pytest.raises(TypeError, match=r"'scale'"):
-            AutoImage(python_img, scale=0.5)
-
-        assert isinstance(AutoImage(python_img), BaseImage)
-
-    def test_from_file(self):
-        with pytest.raises(TypeError, match=r"a string"):
-            from_file(python_img)
-
-        # Ensure size arguments get through
-        with pytest.raises(ValueError, match=r"both width and height"):
-            from_file(python_image, width=1, height=1)
-
-        # Ensure scale argument gets through
-        with pytest.raises(TypeError, match=r"'scale'"):
-            from_file(python_image, scale=1.0)
-
-        assert isinstance(from_file(python_image), BaseImage)
 
 
 class TestCellRatio:
@@ -87,16 +55,3 @@ class TestCellRatio:
             cell_size = (randint(1, 20), randint(1, 20))
             set_cell_size(cell_size)
             assert get_cell_ratio() == truediv(*cell_size) == get_cell_ratio()
-
-
-def test_image_source():
-    assert len(ImageSource) == 3
-    assert all(member.name == name for name, member in ImageSource.__members__.items())
-
-
-def test_style():
-    class MyImage(BaseImage):
-        pass
-
-    assert MyImage.style is None
-    assert str(MyImage) == repr(MyImage)
